+++ src/main/java/ru/rishat/App.java
@@ -2,16 +2,20 @@ package ru.rishat;
 
 import org.openqa.selenium.WebDriver;
 import ru.rishat.config.WebDriverConfig;
+import ru.rishat.controller.FileController;
 import ru.rishat.controller.PositionController;
-import ru.rishat.login.LogIn;
+import ru.rishat.controller.UserController;
+import ru.rishat.entity.User;
 
+import java.io.File;
+import java.io.FileNotFoundException;
 import java.time.Duration;
 import java.time.Instant;
+import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import static ru.rishat.constants.Constants.MARKET_STATE_PLACE;
-import static ru.rishat.constants.Constants.PURCHASE_ID;
+import static ru.rishat.constants.Constants.*;
 
 /**
  *
@@ -20,14 +24,17 @@ public class App {
     private static final Logger logger = Logger.getLogger(App.class.getName());
     private static final WebDriver driver = WebDriverConfig.getWebDriver();
     static PositionController positionController = new PositionController();
+    static UserController userController = new UserController();
+    static FileController fileController = new FileController();
 
-    public static void main(String[] args) throws InterruptedException {
+    public static void main(String[] args) throws InterruptedException, FileNotFoundException {
 
         Instant start = Instant.now();
         logger.log(Level.INFO, "Start program, try scan purchase - " + PURCHASE_ID);
-
         driver.get(MARKET_STATE_PLACE);
-        LogIn.logIn(driver);
+        DATA_FROM_FILE = fileController.readAllLinesFromFile(new File(AUTH_CSV));
+        userController.logInUser(driver, userController.getUser());
+
         synchronized (driver) {
             driver.wait(3000);
         }
