+++ src/main/java/ru/rishat/constants/Constants.java
@@ -1,5 +1,7 @@
 package ru.rishat.constants;
 
+import java.util.List;
+
 public final class Constants {
     public static final long PURCHASE_ID = 36353;
     public static final String MARKET_STATE_PLACE = "https://dedyuhina.posred.pro/purchases/" + PURCHASE_ID + "/market?state=W&place";
@@ -15,7 +17,7 @@ public final class Constants {
     public static final String BOTTOM_TO_PAGE = "//div[@role='status']";
     public static final String PATH_IMAGES_PHOTO_OF_PURCHASE = "src/main/resources/images/" + PURCHASE_ID + "/";
     public static final String AUTH_CSV = "src/main/resources/auth.csv";
-
+    public static List<String> DATA_FROM_FILE;
     public static final String RESOURCES_DATA_XLSX = "src/main/resources/data.xlsx";
     public static final String[] LIST_FOR_VALIDATION_DATA_CELL = {"Куплен", "Не выкуплен", "Ожидает", "Собран", "Получен", "Не найден"};
     public static final String BV = "б/в";
