+++ src/main/java/ru/rishat/controller/FileController.java
@@ -1,0 +1,16 @@
+package ru.rishat.controller;
+
+import ru.rishat.service.FileService;
+import ru.rishat.service.FileServiceImpl;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.util.List;
+
+public class FileController {
+
+    FileService fileService = new FileServiceImpl();
+   public List<String> readAllLinesFromFile(File file) throws FileNotFoundException {
+        return fileService.readAllLinesFromFile(file);
+    }
+}
