+++ src/main/java/ru/rishat/controller/PositionController.java
@@ -1,19 +1,19 @@
 package ru.rishat.controller;
 
 import org.openqa.selenium.WebDriver;
-import ru.rishat.service.PositionService;
+import ru.rishat.service.PositionServiceImp;
 
 import java.util.logging.Logger;
 
 public class PositionController {
     private static final Logger logger = Logger.getLogger(PositionController.class.getName());
-    PositionService positionService = new PositionService();
+    PositionServiceImp positionServiceImp = new PositionServiceImp();
 
     public void scanAllPositions(WebDriver driver) throws InterruptedException {
-        positionService.scanAllPositions(driver);
+        positionServiceImp.scanAllPositions(driver);
     }
 
     public void saveAllPositionsToFile() {
-        positionService.saveAllPositionsToFile();
+        positionServiceImp.saveAllPositionsToFile();
     }
 }
