+++ src/main/java/ru/rishat/dto/PositionDTO.java
@@ -1,0 +1,4 @@
+package ru.rishat.dto;
+
+public class PositionDTO {
+}
