+++ src/main/java/ru/rishat/entity/File.java
@@ -1,0 +1,7 @@
+package ru.rishat.entity;
+
+public class File {
+    private Long id;
+    private String name;
+    private String path;
+}
