+++ src/main/java/ru/rishat/repository/FileRepository.java
@@ -1,0 +1,8 @@
+package ru.rishat.repository;
+
+import java.io.File;
+import java.util.List;
+
+public interface FileRepository {
+    List<String> readAllLinesFromFile(File file);
+}
