+++ src/main/java/ru/rishat/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package ru.rishat.repository;
+
+import ru.rishat.entity.User;
+
+public interface UserRepository {
+    User getUser();
+}
