+++ src/main/java/ru/rishat/service/FileService.java
@@ -1,0 +1,9 @@
+package ru.rishat.service;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.util.List;
+
+public interface FileService {
+    public List<String> readAllLinesFromFile(File file) throws FileNotFoundException;
+}
