+++ src/main/java/ru/rishat/service/UserService.java
@@ -1,0 +1,10 @@
+package ru.rishat.service;
+
+import org.openqa.selenium.WebDriver;
+import ru.rishat.entity.User;
+
+public interface UserService {
+    User getUser();
+
+    void logIn(WebDriver driver, User user);
+}
